Slight refactor or accessors
