+++ src/main/java/com/blamejared/mtlib/utils/BaseAddForestry.java
@@ -1,14 +1,13 @@
-package com.blamejared.mtlib.modspecific;
+package com.blamejared.mtlib.utils;
 
-import com.blamejared.mtlib.utils.BaseUndoable;
 import forestry.api.recipes.ICraftingProvider;
 import forestry.api.recipes.IForestryRecipe;
 
 public abstract class BaseAddForestry<R extends IForestryRecipe> extends BaseUndoable {
-    private ICraftingProvider<R> provider;
+    protected ICraftingProvider<R> provider;
     protected R recipe;
 
-    BaseAddForestry(String name, ICraftingProvider<R> provider, R recipe) {
+    protected BaseAddForestry(String name, ICraftingProvider<R> provider, R recipe) {
         super(name);
         this.provider = provider;
         this.recipe = recipe;
