+++ src/main/java/com/blamejared/mtlib/utils/BaseRemoveForestry.java
@@ -1,6 +1,5 @@
-package com.blamejared.mtlib.modspecific;
+package com.blamejared.mtlib.utils;
 
-import com.blamejared.mtlib.utils.BaseUndoable;
 import forestry.api.recipes.ICraftingProvider;
 import forestry.api.recipes.IForestryRecipe;
 
@@ -8,9 +7,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 public abstract class BaseRemoveForestry<R extends IForestryRecipe> extends BaseUndoable {
-    private ICraftingProvider<R> provider;
+    protected ICraftingProvider<R> provider;
 
-    BaseRemoveForestry(String name, ICraftingProvider<R> provider) {
+    protected BaseRemoveForestry(String name, ICraftingProvider<R> provider) {
         super(name);
         this.provider = provider;
     }
@@ -28,5 +27,5 @@ public abstract class BaseRemoveForestry<R extends IForestryRecipe> extends Base
         }
     }
 
-    abstract boolean checkIsRecipe(R recipe);
+    public abstract boolean checkIsRecipe(R recipe);
 }
